codex parsing: consolidate the duplicate parsers onto the shared modules (message-search F3 phase 2)
The rollout-shape knowledge phase 1 extracted had two more copies; this
lands them on one implementation each, keeping the better body where the
copies had drifted:

- codex_session_file_id: ONE streaming reader in
  external_agent/codex/rollout.rs (the catalog's line-by-line body -
  codex_history's copy read whole files; the finder opens every rollout
  on a cache miss, so streaming matters). Both former homes re-export.
- find_codex_session_file: the catalog's finder now delegates to
  codex_history's engine, which ports the catalog's filename fast path
  (Codex embeds the id in rollout names) IN FRONT of the full scan and
  keeps F5's wrapper-index stored-path cache - so catalog lookups stop
  rescanning migrated homes too (deliberate behavior improvement).
- codex_payload_user_text: reimplemented as the shared
  codex_payload_text shape + role filter + injection filter; the local
  codex_message_content_text duplicate died outright (the compiler
  proved it had no other callers).
- is_codex_injected_user_text_for_main: delegates to the predicate's
  post-phase-1 canonical home instead of routing through the catalog
  shim.
- json_string_field (external_supervision.rs): deleted - it was the
  third value_str clone and the consolidation orphaned its last caller.

Not in scope: reshaping the catalog-typed transcripts.rs reducer arm
onto a typed effective-history API - that lands with the B-wave shard
store, which is its first second consumer.
